+++ Sourcery/Sourcery.swift
@@ -20,7 +20,11 @@ public class Sourcery {
         + "// DO NOT EDIT\n\n"
 
     let verbose: Bool
-    var watcherEnabled: Bool = false
+    private(set) var watcherEnabled: Bool = false
+    private var status = ""
+
+    private var templatesPath: Path = ""
+    private var outputPath: Path = ""
 
     /// Creates Sourcery processor
     ///
@@ -38,6 +42,8 @@ public class Sourcery {
     ///   - watcherEnabled: Whether daemon watcher should be enabled.
     /// - Throws: Potential errors.
     public func processFiles(_ sources: Path, usingTemplates templatesPath: Path, output: Path, watcherEnabled: Bool = false) throws -> [FolderWatcher.Local]? {
+        self.templatesPath = templatesPath
+        self.outputPath = output
         self.watcherEnabled = watcherEnabled
 
         var types: [Type] = try parseTypes(from: sources)
@@ -47,7 +53,7 @@ public class Sourcery {
             return nil
         }
 
-        print("Starting watching sources.")
+        track("Starting watching sources.", skipStatus: true)
         let sourceWatcher = FolderWatcher.Local(path: sources.string) { events in
             let events = events
                 .filter { $0.flags.contains(.isFile) && $0.path.hasSuffix(".swift") }
@@ -63,16 +69,16 @@ public class Sourcery {
 
             if shouldRegenerate {
                 do {
-                    print("Source changed: ", terminator: "")
+                    self.track("Source changed: ", terminator: "")
                     types = try self.parseTypes(from: sources)
                     _ = try self.generate(templatePath: templatesPath, output: output, types: types)
                 } catch {
-                    print(error)
+                    self.track(error)
                 }
             }
         }
 
-        print("Starting watching templates.")
+        track("Starting watching templates.", skipStatus: true)
 
         let templateWatcher = FolderWatcher.Local(path: templatesPath.string) { events in
             let events = events
@@ -80,10 +86,10 @@ public class Sourcery {
 
             if !events.isEmpty {
                 do {
-                    print("Template changed: ", terminator: "")
+                    self.track("Templates changed: ", terminator: "")
                     _ = try self.generate(templatePath: templatesPath, output: output, types: types)
                 } catch {
-                    print(error)
+                    self.track(error)
                 }
             }
         }
@@ -92,7 +98,7 @@ public class Sourcery {
     }
 
     private func parseTypes(from: Path) throws -> [Type] {
-        print("Scanning sources...", terminator: "")
+        self.track("Scanning sources...", terminator: "")
         let parser = Parser(verbose: verbose)
 
         var parserResult: ParserResult = ([], [])
@@ -102,28 +108,37 @@ public class Sourcery {
             return parser.uniqueTypes(parserResult)
         }
 
-        try from
+        let sources = try from
             .recursiveChildren()
             .filter {
                 $0.extension == "swift"
             }
-            .forEach { path in
+
+        var lastIdx = 0
+        let step = sources.count / 10 // every 10%
+        try sources.enumerated().forEach { idx, path in
+                if idx > lastIdx + step {
+                    lastIdx = idx
+                    let percentage = idx * 100 / sources.count
+                    self.track("Scanning sources... \(percentage)% (\(sources.count) files)", terminator: "")
+                }
                 parserResult = try parser.parseFile(path, existingTypes: parserResult)
         }
 
         //! All files have been scanned, time to join extensions with base class
         let types = parser.uniqueTypes(parserResult)
 
-        print("Found \(types.count) types.")
+        track("Found \(types.count) types.")
         return types
     }
 
     private func generate(templatePath: Path, output: Path, types: [Type]) throws {
-        print("Loading templates...", terminator: "")
+        track("Loading templates...", terminator: "")
         let allTemplates = try templates(from: templatePath)
-        print("Loaded \(allTemplates.count) templates.")
+        track("Loaded \(allTemplates.count) templates.")
 
-        print("Generating code...", terminator: "")
+        track("Generating code...", terminator: "")
+        status = ""
 
         guard output.isDirectory else {
             let result = try allTemplates.reduce(Sourcery.generationHeader) { result, template in
@@ -140,12 +155,11 @@ public class Sourcery {
             try outputPath.write(result, encoding: .utf8)
         }
 
-        print("Finished.")
+        track("Finished.", skipStatus: true)
     }
 
     private func generate(_ template: Template, forTypes types: [Type]) throws -> String {
-        let shouldRecover = watcherEnabled
-        guard shouldRecover else {
+        guard watcherEnabled else {
             return try Generator.generate(types, template: template)
         }
 
@@ -155,6 +169,7 @@ public class Sourcery {
         }, catch: { error in
             result = error?.description ?? ""
         }, finallyBlock: {})
+
         return result
     }
 
@@ -163,17 +178,37 @@ public class Sourcery {
     }
 
     private func templates(from: Path) throws -> [SourceryTemplate] {
+        return try templatePaths(from: from).map {
+                try SourceryTemplate(path: $0)
+        }
+    }
+
+    private func outputPaths(from: Path, output: Path) throws -> [Path] {
+        return try templatePaths(from: from).map { output + generatedPath(for: $0) }
+    }
+
+    private func templatePaths(from: Path) throws -> [Path] {
         guard from.isDirectory else {
-            return [try SourceryTemplate(path: from)]
+            return [from]
         }
 
         return try from
             .recursiveChildren()
             .filter {
                 $0.extension == "stencil"
-            }
-            .map {
-                try SourceryTemplate(path: $0)
+        }
+    }
+
+    private func track(_ message: Any, terminator: String = "\n", skipStatus: Bool = false) {
+        if !watcherEnabled || verbose {
+            Swift.print(message, terminator: terminator)
+        }
+
+        guard watcherEnabled && !skipStatus else { return }
+        status = String(describing: message) + terminator
+
+        _ = try? outputPaths(from: templatesPath, output: outputPath).forEach { path in
+            _ = try? path.write(Sourcery.generationHeader + "STATUS:\n" + status, encoding: .utf8)
         }
     }
 }